Update releasing.md with Kubernetes release process
This updates releasing.md with actual instructions on how to cut a
release, leaving the theory section of that document alone. Along the
way, I streamlined tiny bits of the existing process as I was
describing them.

The instructions are possibly pedantic, but should be executable by
anyone at this point, versus taking someone versant in the dark arts.

Relies on #10910. Fixes #1883.